+++ src/pydiverse/pipedag/backend/table/sql/dialects/mssql.py
@@ -1,4 +1,4 @@
-# Copyright (c) QuantCo and pydiverse contributors 2025-2025
+# Copyright (c) QuantCo and pydiverse contributors 2025-2026
 # SPDX-License-Identifier: BSD-3-Clause
 import copy
 import dataclasses
@@ -22,6 +22,7 @@ from pydiverse.pipedag.backend.table.sql.ddl import (
     AddClusteredColumnstoreIndex,
     ChangeColumnTypes,
     CreateAlias,
+    DropTable,
     _mssql_update_definition,
 )
 from pydiverse.pipedag.backend.table.sql.reflection import PipedagMSSqlReflection
@@ -435,10 +436,13 @@ class DataframeMsSQLTableHook:
             )
 
         if mssqlkit:
+            if not early:
+                cls._dialect_create_empty_table(store, table, schema, dtypes)
             mss.table.bulk_upload(
                 engine=store.engine,
                 table_name=f"{schema_name}.{table.name}",
                 data=df,
+                string_encoding="cp1252",  # MSSQL ODBC is more reliable with 1-byte encodings
             )
         else:
             # use bcpandas
@@ -499,7 +503,7 @@ class DataframeMsSQLTableHook:
         cls,
         query: Any,
         store: SQLTableStore,
-        dtypes: dict[str, pl.DataFrame] | None = None,
+        dtypes: dict[str, pl.DataType] | None = None,
     ) -> pl.DataFrame:
         assert dtypes is None, (
             "Pyarrow reads SQL schema and loads the data in reasonable types."
@@ -589,7 +593,7 @@ class PandasTableHook(DataframeMsSQLTableHook, sql_hooks.PandasTableHook):
                 df = pl_df.to_pandas()
                 return cls._fix_dtypes(df, dtypes)
             except Exception as e:  # noqa
-                store.logger.warning("Failed to download table using arrow-odbc, falling back to sqlalchemy/pandas.")
+                store.logger.exception("Failed to download table using arrow-odbc, falling back to sqlalchemy/pandas.")
 
         return super().download_table(query, store, dtypes)
 
@@ -607,6 +611,7 @@ class PandasTableHook(DataframeMsSQLTableHook, sql_hooks.PandasTableHook):
                 return cls.upload_table_bulk_insert(table, schema, dtypes, store, early)
             except Exception as e:  # noqa
                 store.logger.exception("Failed to upload table using bulk insert, falling back to pandas.")
+                store.execute(DropTable(table.name, schema, if_exists=True, cascade=True))
         # TODO: consider using arrow-odbc for uploading
         super().upload_table(table, schema, dtypes, store, early)
 
@@ -638,7 +643,7 @@ class PolarsTableHook(DataframeMsSQLTableHook, sql_hooks.PolarsTableHook):
         cls,
         query: Any,
         store: SQLTableStore,
-        dtypes: dict[str, pl.DataFrame] | None = None,
+        dtypes: dict[str, pl.DataType] | None = None,
     ) -> pl.DataFrame:
         assert dtypes is None, (
             "Pyarrow reads SQL schema and loads the data in reasonable types."
@@ -651,7 +656,7 @@ class PolarsTableHook(DataframeMsSQLTableHook, sql_hooks.PolarsTableHook):
             try:
                 return cls.download_table_arrow_odbc(query, store, dtypes)
             except Exception as e:  # noqa
-                store.logger.warning(
+                store.logger.exception(
                     "Failed to download table using arrow-odbc, falling back to "
                     "sqlalchemy/polars which currently uses pandas in the back."
                 )
@@ -675,7 +680,10 @@ class PolarsTableHook(DataframeMsSQLTableHook, sql_hooks.PolarsTableHook):
             try:
                 return cls.upload_table_bulk_insert(table, schema, dtypes, store, early)
             except:  # noqa
-                store.logger.warning("Failed to upload table using bulk insert, falling back to polars.write_database.")
+                store.logger.exception(
+                    "Failed to upload table using bulk insert, falling back to polars.write_database."
+                )
+                store.execute(DropTable(table.name, schema, if_exists=True, cascade=True))
         super().upload_table(table, schema, dtypes, store, early)