+++ docs/source/changelog.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## 0.12.11 (2026-01-21)
+## 0.12.11 (2026-01-23)
 - Feat: simplify accessing tables in ParquetTableStore with polars
+- Fix: Fix mssqlkit bulk upload
+- Speed up pandas/polars transfer to/from duckdb for SqlTableStore
 
 ## 0.12.10 (2026-01-20)
 - Feat: Add ParquetTableStore.sync_metadata() function to sync local duckdb file with metadata store without running